header logic working
