--- src/components/TicketModal/components/EditableText/styles.module.scss
@@ -1,4 +1,0 @@
-.text-box {
-    color: white;
-    border: 0;
-}
