--- src/components/TicketModal/components/StatusDropdown/styles.module.scss
@@ -1,3 +1,0 @@
-.dropdown-item:hover {
-    cursor:pointer;
-}
